[FTP] fix compilation errors

+++ ftp/src/common.d/ft_init.c
@@ -6,16 +6,17 @@
 /*   By: gbourgeo <gbourgeo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/17 04:08:22 by gbourgeo          #+#    #+#             */
-/*   Updated: 2019/10/21 00:09:34 by gbourgeo         ###   ########.fr       */
+/*   Updated: 2019/10/24 20:43:07 by gbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include <unistd.h>
 #include "libft.h"
 #include "common.h"
 
 extern char		**environ;
 
-int				ft_init(void *p, size_t size, int type, char *prog)
+int				ft_init(void *p, int size, int type, char *prog)
 {
 	t_common	*c;
